# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Styling with Muskan API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/request-otp": {
      "post": {
        "summary": "Request OTP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "phone": { "type": "string" } }, "required": ["phone"] }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP sent" },
          "400": { "description": "Validation errors" }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "summary": "Verify OTP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "phone": { "type": "string" }, "otp": { "type": "string" } }, "required": ["phone", "otp"] }
            }
          }
        },
        "responses": {
          "200": { "description": "Authenticated" },
          "400": { "description": "Invalid OTP or validation errors" }
        }
      }
    },
...
    "/content/categories": {
      "get": {
        "summary": "List categories",
        "parameters": [{ "name": "gender", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/content/services": {
      "get": {
        "summary": "List services",
        "parameters": [{ "name": "category", "in": "query", "schema": { "type": "string" } }, { "name": "gender", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}